+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/model/KafkaMirrorMaker2Cluster.java
@@ -7,11 +7,15 @@ package io.strimzi.operator.cluster.model;
 import io.fabric8.kubernetes.api.model.EnvVar;
 import io.fabric8.kubernetes.api.model.Volume;
 import io.fabric8.kubernetes.api.model.VolumeMount;
+import io.fabric8.kubernetes.api.model.rbac.PolicyRule;
+import io.fabric8.kubernetes.api.model.rbac.PolicyRuleBuilder;
+import io.fabric8.kubernetes.api.model.rbac.Role;
 import io.strimzi.api.kafka.model.common.CertSecretSource;
 import io.strimzi.api.kafka.model.common.ClientTls;
 import io.strimzi.api.kafka.model.common.authentication.KafkaClientAuthentication;
 import io.strimzi.api.kafka.model.common.authentication.KafkaClientAuthenticationOAuth;
 import io.strimzi.api.kafka.model.common.authentication.KafkaClientAuthenticationTls;
+import io.strimzi.api.kafka.model.connect.KafkaConnectResources;
 import io.strimzi.api.kafka.model.connect.KafkaConnectSpec;
 import io.strimzi.api.kafka.model.connect.KafkaConnectSpecBuilder;
 import io.strimzi.api.kafka.model.mirrormaker2.KafkaMirrorMaker2;
@@ -44,12 +48,6 @@ public class KafkaMirrorMaker2Cluster extends KafkaConnectCluster {
     protected static final String COMPONENT_TYPE = "kafka-mirror-maker-2";
 
     // Kafka MirrorMaker 2 connector configuration keys (EnvVariables)
-    protected static final String ENV_VAR_KAFKA_MIRRORMAKER_2_CLUSTERS = "KAFKA_MIRRORMAKER_2_CLUSTERS";
-    protected static final String ENV_VAR_KAFKA_MIRRORMAKER_2_TRUSTED_CERTS_CLUSTERS = "KAFKA_MIRRORMAKER_2_TRUSTED_CERTS_CLUSTERS";
-    protected static final String ENV_VAR_KAFKA_MIRRORMAKER_2_TLS_AUTH_CLUSTERS = "KAFKA_MIRRORMAKER_2_TLS_AUTH_CLUSTERS";
-    protected static final String ENV_VAR_KAFKA_MIRRORMAKER_2_TLS_AUTH_CERTS_CLUSTERS = "KAFKA_MIRRORMAKER_2_TLS_AUTH_CERTS_CLUSTERS";
-    protected static final String ENV_VAR_KAFKA_MIRRORMAKER_2_TLS_AUTH_KEYS_CLUSTERS = "KAFKA_MIRRORMAKER_2_TLS_AUTH_KEYS_CLUSTERS";
-    protected static final String ENV_VAR_KAFKA_MIRRORMAKER_2_OAUTH_TRUSTED_CERTS_CLUSTERS = "KAFKA_MIRRORMAKER_2_OAUTH_TRUSTED_CERTS_CLUSTERS";
     protected static final String CO_ENV_VAR_CUSTOM_MIRROR_MAKER2_POD_LABELS = "STRIMZI_CUSTOM_KAFKA_MIRROR_MAKER2_LABELS";
 
     protected static final String MIRRORMAKER_2_OAUTH_SECRETS_BASE_VOLUME_MOUNT = "/opt/kafka/mm2-oauth/";
@@ -287,24 +285,19 @@ public class KafkaMirrorMaker2Cluster extends KafkaConnectCluster {
         return KafkaMirrorMaker2Resources.mm2RoleBindingName(getCluster());
     }
 
+    // TODO: We may not need this method at all because all client auth configurations will use config provider instead
     @Override
     protected List<Volume> getVolumes(boolean isOpenShift) {
         List<Volume> volumeList = super.getVolumes(isOpenShift);
 
         for (KafkaMirrorMaker2ClusterSpec mirrorMaker2Cluster: clusters) {
-            String alias = mirrorMaker2Cluster.getAlias();
-            ClientTls tls = mirrorMaker2Cluster.getTls();
-
-            if (tls != null) {
-                CertUtils.createTrustedCertificatesVolumes(volumeList, tls.getTrustedCertificates(), isOpenShift, alias);
-            }
-
-            AuthenticationUtils.configurePKCS12ClientAuthenticationVolumes(mirrorMaker2Cluster.getAuthentication(), volumeList, mirrorMaker2Cluster.getAlias() + "-oauth-certs", isOpenShift, mirrorMaker2Cluster.getAlias() + '-',  true);
+            AuthenticationUtils.configureClientAuthenticationVolumes(mirrorMaker2Cluster.getAuthentication(), volumeList, KafkaConnectResources.internalOauthTrustedCertsSecretName(cluster), isOpenShift, mirrorMaker2Cluster.getAlias() + "-", true);
         }
         return volumeList;
     }
 
     @Override
+    // TODO: We may not need this method at all because all client auth configurations will use config provider instead
     protected List<VolumeMount> getVolumeMounts() {
         List<VolumeMount> volumeMountList = super.getVolumeMounts();
 
@@ -312,17 +305,41 @@ public class KafkaMirrorMaker2Cluster extends KafkaConnectCluster {
             String alias = mirrorMaker2Cluster.getAlias();
             String tlsVolumeMountPath =  buildClusterVolumeMountPath(MIRRORMAKER_2_TLS_CERTS_BASE_VOLUME_MOUNT, alias);
 
-            ClientTls kafkaMirrorMaker2Tls = mirrorMaker2Cluster.getTls();
-            if (kafkaMirrorMaker2Tls != null) {
-                CertUtils.createTrustedCertificatesVolumeMounts(volumeMountList, kafkaMirrorMaker2Tls.getTrustedCertificates(), tlsVolumeMountPath, alias);
-            }
             String passwordVolumeMountPath =  buildClusterVolumeMountPath(MIRRORMAKER_2_PASSWORD_VOLUME_MOUNT, alias);
             String oauthTlsVolumeMountPath =  buildClusterVolumeMountPath(MIRRORMAKER_2_OAUTH_TLS_CERTS_BASE_VOLUME_MOUNT, alias);
             String oauthVolumeMountPath =  buildClusterVolumeMountPath(MIRRORMAKER_2_OAUTH_SECRETS_BASE_VOLUME_MOUNT, alias);
-            AuthenticationUtils.configurePKCS12ClientAuthenticationVolumeMounts(mirrorMaker2Cluster.getAuthentication(), volumeMountList, tlsVolumeMountPath, passwordVolumeMountPath, oauthTlsVolumeMountPath, mirrorMaker2Cluster.getAlias() + "-oauth-certs", mirrorMaker2Cluster.getAlias() + '-', true, oauthVolumeMountPath);
+            AuthenticationUtils.configureClientAuthenticationVolumeMounts(mirrorMaker2Cluster.getAuthentication(), volumeMountList, tlsVolumeMountPath, passwordVolumeMountPath, oauthTlsVolumeMountPath, KafkaConnectResources.internalOauthTrustedCertsSecretName(cluster), mirrorMaker2Cluster.getAlias() + '-', true, oauthVolumeMountPath);
         }
 
         return volumeMountList;
+    }
+    @Override
+    public Role generateRole() {
+        List<String> certSecretNames = new ArrayList<>();
+        for (KafkaMirrorMaker2ClusterSpec mirrorMaker2Cluster: clusters) {
+            ClientTls mirrorTls = mirrorMaker2Cluster.getTls();
+            if (mirrorTls != null && mirrorTls.getTrustedCertificates() != null && !mirrorTls.getTrustedCertificates().isEmpty()) {
+                certSecretNames.add(KafkaConnectResources.internalTlsTrustedCertsSecretName(cluster));
+            }
+
+            KafkaClientAuthentication mirrorAuthentication = mirrorMaker2Cluster.getAuthentication();
+
+            if (mirrorAuthentication instanceof KafkaClientAuthenticationTls tlsAuth && tlsAuth.getCertificateAndKey() != null) {
+                certSecretNames.add(tlsAuth.getCertificateAndKey().getSecretName());
+            } else if (mirrorAuthentication instanceof KafkaClientAuthenticationOAuth oauth && oauth.getTlsTrustedCertificates() != null
+                    && !oauth.getTlsTrustedCertificates().isEmpty()) {
+                certSecretNames.add(KafkaConnectResources.internalOauthTrustedCertsSecretName(cluster));
+            }
+        }
+
+        List<PolicyRule> rules = List.of(new PolicyRuleBuilder()
+                .withApiGroups("")
+                .withResources("secrets")
+                .withVerbs("get")
+                .withResourceNames(certSecretNames)
+                .build());
+
+        return RbacUtils.createRole(componentName, namespace, rules, labels, ownerReference, null);
     }
 
     private static String buildClusterVolumeMountPath(final String baseVolumeMount,  final String path) {
@@ -333,53 +350,6 @@ public class KafkaMirrorMaker2Cluster extends KafkaConnectCluster {
     @Override
     protected List<EnvVar> getEnvVars() {
         List<EnvVar> varList = super.getEnvVars();
-
-        final StringBuilder clusterAliases = new StringBuilder();
-        final StringBuilder clustersTrustedCerts = new StringBuilder();
-        final StringBuilder clustersTlsAuthCerts = new StringBuilder();
-        final StringBuilder clustersTlsAuthKeys = new StringBuilder();
-        final StringBuilder clustersOauthTrustedCerts = new StringBuilder();
-
-        for (KafkaMirrorMaker2ClusterSpec mirrorMaker2Cluster : clusters) {
-            String clusterAlias = mirrorMaker2Cluster.getAlias();
-
-            if (clusterAliases.length() > 0) {
-                clusterAliases.append(";");
-            }
-            clusterAliases.append(clusterAlias);
-
-            getClusterTrustedCerts(clustersTrustedCerts, mirrorMaker2Cluster, clusterAlias);
-
-            KafkaClientAuthentication authentication = mirrorMaker2Cluster.getAuthentication();
-            if (authentication != null) {
-                if (authentication instanceof KafkaClientAuthenticationTls tlsAuth) {
-                    if (tlsAuth.getCertificateAndKey() != null) {
-                        appendCluster(clustersTlsAuthCerts, clusterAlias, () -> tlsAuth.getCertificateAndKey().getSecretName() + "/" + tlsAuth.getCertificateAndKey().getCertificate());
-                        appendCluster(clustersTlsAuthKeys, clusterAlias, () -> tlsAuth.getCertificateAndKey().getSecretName() + "/" + tlsAuth.getCertificateAndKey().getKey());
-                    }
-                } else if (authentication instanceof KafkaClientAuthenticationOAuth oauth) {
-                    if (oauth.getTlsTrustedCertificates() != null && !oauth.getTlsTrustedCertificates().isEmpty()) {
-                        appendClusterOAuthTrustedCerts(clustersOauthTrustedCerts, clusterAlias, oauth.getTlsTrustedCertificates());
-                    }
-                }
-            }
-        }
-
-        varList.add(ContainerUtils.createEnvVar(ENV_VAR_KAFKA_MIRRORMAKER_2_CLUSTERS, clusterAliases.toString()));
-
-        if (clustersTrustedCerts.length() > 0) {
-            varList.add(ContainerUtils.createEnvVar(ENV_VAR_KAFKA_MIRRORMAKER_2_TRUSTED_CERTS_CLUSTERS, clustersTrustedCerts.toString()));
-        }
-
-        if (clustersTlsAuthCerts.length() > 0 || clustersTlsAuthKeys.length() > 0) {
-            varList.add(ContainerUtils.createEnvVar(ENV_VAR_KAFKA_MIRRORMAKER_2_TLS_AUTH_CLUSTERS, "true"));
-            varList.add(ContainerUtils.createEnvVar(ENV_VAR_KAFKA_MIRRORMAKER_2_TLS_AUTH_CERTS_CLUSTERS, clustersTlsAuthCerts.toString()));
-            varList.add(ContainerUtils.createEnvVar(ENV_VAR_KAFKA_MIRRORMAKER_2_TLS_AUTH_KEYS_CLUSTERS, clustersTlsAuthKeys.toString()));
-        }
-
-        if (clustersOauthTrustedCerts.length() > 0) {
-            varList.add(ContainerUtils.createEnvVar(ENV_VAR_KAFKA_MIRRORMAKER_2_OAUTH_TRUSTED_CERTS_CLUSTERS, clustersOauthTrustedCerts.toString()));
-        }
 
         JvmOptionUtils.jvmSystemProperties(varList, jvmOptions);
 
@@ -400,6 +370,7 @@ public class KafkaMirrorMaker2Cluster extends KafkaConnectCluster {
                 clustersTrustedCerts.append(CertUtils.trustedCertsEnvVar(trustedCertificates));
             }
         }
+        return secretsToCopy;
     }
 
     private static void appendClusterOAuthTrustedCerts(final StringBuilder clusters, String clusterAlias, List<CertSecretSource> trustedCerts) {